+++ workspaces/azure-sites/plugins/azure-sites-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-azure-sites-common
 
+## 0.1.8
+
+### Patch Changes
+
+- 58c28d3: Backstage version bump to v1.32.2
+
 ## 0.1.7
 
 ### Patch Changes