+++ workspaces/bazaar/plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-bazaar
 
+## 0.5.1
+
+### Patch Changes
+
+- 56ae2fe: Backstage version bump to v1.32.2
+
+## 0.5.0
+
+### Minor Changes
+
+- fe0d104: Added a new UserSelector component that allows users to select a responsible user from the catalog, with the option to manually enter still being there"
+
 ## 0.4.0
 
 ### Minor Changes