+++ workspaces/copilot/plugins/copilot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage-community/plugin-copilot
 
+## 0.2.3
+
+### Patch Changes
+
+- 399dc3b: Backstage version bump to v1.32.2
+- Updated dependencies [399dc3b]
+  - @backstage-community/plugin-copilot-common@0.2.2
+
 ## 0.2.2
 
 ### Patch Changes