+++ workspaces/xcmetrics/plugins/xcmetrics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-xcmetrics
 
+## 0.2.58
+
+### Patch Changes
+
+- 0f0bc98: Backstage version bump to v1.32.2
+
 ## 0.2.57
 
 ### Patch Changes